backend/requirements.txt
fastapi==0.117.1
uvicorn[standard]==0.36.0
pillow==11.2.1
python-multipart==0.0.20
